Rename section -> accordion
