+++ packages/11ty/_plugins/shortcodes/index.js
@@ -1,3 +1,4 @@
+const accordion = require('./accordion.js')
 const addComponentTag = require('../../_plugins/components/addComponentTag')
 const annoref = require('./annoref')
 const backmatter = require('./backmatter')
@@ -7,7 +8,6 @@ const contributors = require('./contributors')
 const figure = require('./figure')
 const figureGroup = require('./figureGroup')
 const ref = require('./figureRef')
-const section = require('./section.js')
 const shortcodeFactory = require('../components/shortcodeFactory')
 const title = require('./title')
 const tombstone = require('./tombstone')
@@ -15,9 +15,9 @@ const tombstone = require('./tombstone')
 module.exports = function(eleventyConfig, collections, options) {
   const { addShortcode, addPairedShortcode } = shortcodeFactory(eleventyConfig, collections)
 
+  addPairedShortcode('accordion', accordion)
   addComponentTag(eleventyConfig, 'annoref', annoref)
   addPairedShortcode('backmatter', backmatter)
-  addPairedShortcode('section', section)
   addShortcode('bibliography', bibliography)
   addShortcode('cite', cite)
   addComponentTag(eleventyConfig, 'contributors', contributors)
